pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rares Matei <rares.matei@sap.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
pdfminer-six = "^20250506"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
sentence-transformers = "^4.1.0"
faiss-cpu = "^1.11.0"
python-dotenv = "^1.1.0"
pypdf = "^5.5.0"
pinecone = "^7.0.0"
pyreadline3 = "^3.5.4"
openai = "^1.81.0"
pytest = "^8.3.5"
pydantic-settings = "^2.9.1"
mangum = "^0.19.0"
authlib = "^1.6.0"
fastapi-jwt-auth = "^0.5.0"
sqlalchemy = "^2.0.41"
psycopg = {extras = ["binary"], version = "^3.2.9"}
alembic = "^1.16.1"
passlib = "^1.7.4"
python-jwt = "^4.1.0"
psycopg2 = "^2.9.10"
asyncpg = "^0.30.0"
bcrypt = "^4.3.0"
watchfiles = "^1.0.5"
hypercorn = "^0.17.3"

[tool.poetry.scripts]
scrape-legislation = "scrape_api.cli:main"
ingest             = "scrape_api.ingest:main"

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
